refactor(PatternViewer): Extract PatternViewer from RouteViewer.
