+++ lib/components/viewers/route-viewer.js
@@ -4,7 +4,6 @@ import { Button } from 'react-bootstrap'
 import { connect } from 'react-redux'
 import { Filter } from '@styled-icons/fa-solid/Filter'
 import { FormattedMessage, injectIntl } from 'react-intl'
-import { getMostReadableTextColor } from '@opentripplanner/core-utils/lib/route'
 import { Search } from '@styled-icons/fa-solid/Search'
 import coreUtils from '@opentripplanner/core-utils'
 import PropTypes from 'prop-types'
@@ -12,25 +11,17 @@ import React, { Component } from 'react'
 
 import * as apiActions from '../../actions/api'
 import * as uiActions from '../../actions/ui'
-import { ComponentContext } from '../../util/contexts'
-import {
-  extractHeadsignFromPattern,
-  getModeFromRoute,
-  getRouteColorBasedOnSettings
-} from '../../util/viewer'
 import {
   getAgenciesFromRoutes,
   getModesForActiveAgencyFilter,
   getOperatorAndRoute,
-  getRouteOperator,
   getSortedFilteredRoutes
 } from '../../util/state'
 import { getFormattedMode } from '../../util/i18n'
 import { StyledIconWrapper } from '../util/styledIcon'
 import PageTitle from '../util/page-title'
 
-import { RouteName, RouteRow } from './RouteRow'
-import RouteDetails from './route-details'
+import { RouteRow } from './RouteRow'
 
 class RouteViewer extends Component {
   static propTypes = {
@@ -59,45 +50,25 @@ class RouteViewer extends Component {
     initialRender: true
   }
 
-  static contextType = ComponentContext
-
   /**
-   * If we're viewing a pattern's stops, route to
-   * main route viewer, otherwise go back to main view
+   * Route back to main view.
    */
-  _backClicked = () =>
-    this.props.viewedRoute === null
-      ? this.props.setMainPanelContent(null)
-      : this.props.setViewedRoute({
-          ...this.props.viewedRoute,
-          patternId: null
-        })
+  _backClicked = () => this.props.setMainPanelContent(null)
 
   /**
    * Gets a breadcrumbs-like title following these rules,
    * so we don't need to internationalize the title bar structure:
    * - if no route is selected, just say "Route Viewer",
-   * - if a route is selected, display the agency and user-facing route number,
-   * - if a pattern is selected, add the pattern destination too.
+   * - if a route is selected, display the agency and user-facing route number.
    */
   getTitle = () => {
-    const { intl, transitOperators, viewedRoute, viewedRouteObject } =
-      this.props
-    const { patternId } = viewedRoute || {}
-    const { patterns, pending, shortName } = viewedRouteObject || {}
+    const { intl, transitOperators, viewedRouteObject } = this.props
+    const { pending } = viewedRouteObject || {}
     if (!viewedRouteObject || pending) {
       return intl.formatMessage({ id: 'components.RouteViewer.title' })
     }
 
-    const pattern = patterns?.[patternId]
-    return (
-      getOperatorAndRoute(viewedRouteObject, transitOperators, intl) +
-      (patternId && pattern
-        ? ` ${intl.formatMessage({
-            id: 'components.RouteDetails.stopsTo'
-          })} ${extractHeadsignFromPattern(pattern, shortName)}`
-        : '')
-    )
+    return getOperatorAndRoute(viewedRouteObject, transitOperators, intl)
   }
 
   componentDidMount() {
@@ -158,68 +129,10 @@ class RouteViewer extends Component {
       routes: sortedRoutes,
       setViewedRoute,
       transitOperators,
-      vehicleIconHighlight,
-      viewedRoute,
-      viewedRouteObject
+      viewedRoute
     } = this.props
 
     const { initialRender } = this.state
-    const { ModeIcon, RouteRenderer } = this.context
-
-    // If patternId is present, we're looking at a specific pattern's stops
-    if (viewedRoute?.patternId && viewedRouteObject) {
-      const { patternId } = viewedRoute
-      const route = viewedRouteObject
-      // Find operator based on agency_id (extracted from OTP route ID).
-      const operator = getRouteOperator(viewedRouteObject, transitOperators)
-      const routeColor = getRouteColorBasedOnSettings(operator, route)
-      const textColor = getMostReadableTextColor(routeColor, route?.textColor)
-      const fill = vehicleIconHighlight === false ? null : textColor
-
-      return (
-        <div
-          className="route-viewer"
-          style={{
-            backgroundColor: routeColor,
-            color: textColor,
-            fill
-          }}
-        >
-          <PageTitle title={this.getTitle()} />
-          {/* Header Block */}
-          <div
-            className="route-viewer-header"
-            style={{ backgroundColor: routeColor }}
-          >
-            <div className="header-text route-expanded">
-              <h1 style={{ display: 'contents' }}>
-                {!route.pending && ModeIcon && (
-                  <ModeIcon
-                    aria-label={getFormattedMode(
-                      getModeFromRoute(route).toLowerCase(),
-                      intl
-                    )}
-                    mode={getModeFromRoute(route)}
-                    style={{ maxHeight: 40 }}
-                    width={22}
-                  />
-                )}
-                <RouteName
-                  basicRender
-                  route={route}
-                  RouteRenderer={RouteRenderer}
-                />
-              </h1>
-            </div>
-          </div>
-          <RouteDetails
-            operator={operator}
-            pattern={route?.patterns?.[patternId]}
-            route={route}
-          />
-        </div>
-      )
-    }
     const { search } = filter
     const operators =
       transitOperators.length > 0
@@ -354,7 +267,6 @@ const mapStateToProps = (state) => {
     modes: getModesForActiveAgencyFilter(state),
     routes: getSortedFilteredRoutes(state),
     transitOperators: state.otp.config.transitOperators,
-    vehicleIconHighlight: state.otp.config?.routeViewer?.vehicleIconHighlight,
     viewedRoute: state.otp.ui.viewedRoute,
     viewedRouteObject:
       state.otp.transitIndex.routes?.[state.otp.ui.viewedRoute?.routeId]
