+++ lib/components/viewers/viewer-container.js
@@ -4,6 +4,7 @@ import React, { Component } from 'react'
 
 import { MainPanelContent } from '../../actions/ui-constants'
 
+import PatternViewer from './pattern-viewer'
 import RouteViewer from './route-viewer'
 import StopViewer from './stop-viewer'
 import TripViewer from './trip-viewer'
@@ -20,11 +21,10 @@ class ViewerContainer extends Component {
     const { className, style, uiState } = this.props
 
     // check for main panel content
-    if (
-      uiState.mainPanelContent === MainPanelContent.ROUTE_VIEWER ||
-      uiState.mainPanelContent === MainPanelContent.PATTERN_VIEWER
-    ) {
+    if (uiState.mainPanelContent === MainPanelContent.ROUTE_VIEWER) {
       return <RouteViewer hideBackButton />
+    } else if (uiState.mainPanelContent === MainPanelContent.PATTERN_VIEWER) {
+      return <PatternViewer hideBackButton />
     }
 
     // check for stop viewer
@@ -32,11 +32,10 @@ class ViewerContainer extends Component {
       return <StopViewer hideBackButton />
     }
 
+    // check for trip viewer
     if (uiState.mainPanelContent === MainPanelContent.TRIP_VIEWER) {
       return <TripViewer hideBackButton />
     }
-
-    // check for trip viewer
 
     // otherwise, return default content
     return (
